requirements.txt
python-telegram-bot[job-queue]==21.1.1
httpx==0.27.0
python-dotenv==1.0.1
aiosqlite==0.20.0
fastapi==0.111.0
uvicorn==0.29.0
tenacity==8.3.0